Moved custom color class to CodeBlockUtils
