+++ src/DynamoCore/Utilities/CodeBlockUtils.cs
@@ -3,6 +3,7 @@ using Dynamo.Nodes;
 using Dynamo.UI;
 using Dynamo.UI.Controls;
 using ICSharpCode.AvalonEdit.Highlighting;
+using ICSharpCode.AvalonEdit.Rendering;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
@@ -338,4 +339,33 @@ namespace Dynamo.Utilities
             return digitRule;
         }
     }
+
+    // Refer to link: 
+    // http://stackoverflow.com/questions/11806764/adding-syntax-highlighting-rules-to-avalonedit-programmatically
+    internal sealed class CustomizedBrush : HighlightingBrush
+    {
+        private readonly SolidColorBrush brush;
+        public CustomizedBrush(Color color)
+        {
+            brush = CreateFrozenBrush(color);
+        }
+
+        public override Brush GetBrush(ITextRunConstructionContext context)
+        {
+            return brush;
+        }
+
+        public override string ToString()
+        {
+            return brush.ToString();
+        }
+
+        private static SolidColorBrush CreateFrozenBrush(Color color)
+        {
+            SolidColorBrush brush = new SolidColorBrush(color);
+            brush.Freeze();
+            return brush;
+        }
+    }
+
 }
